more general block action
